+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 1.0.0-beta.2 - 1/14/2017
 
+- Update AppComponent to be `<app-root />` to satisfy TSLint. `bootstrap-server.ts` updated as well.
+
 - Partial revert of ng2-bootstrap dependency down to **1.1.16**, 1.1.20+ require ng 2.4.x (unsupported by Universal atm) [commit](https://github.com/MarkPieszak/aspnetcore-angular2-universal/commit/524df2df00113d0ee2953b44ae40167112192f89)
 
 - [#33](https://github.com/MarkPieszak/aspnetcore-angular2-universal/issues/33) Update CSS for bootstrap4 & fix ng2-bootstrap API changes  [commit](https://github.com/MarkPieszak/aspnetcore-angular2-universal/commit/d0c0e7d98b9ac043be9880ba2656ddf0f0f2222d)